requirements.txt
asgiref==3.7.2
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.2.0
django-extra-fields==3.0.2
djangorestframework==3.14.0
dnspython==2.3.0
filelock==3.11.0
Flask==2.3.2
fpdf==1.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.1.1
Pillow==10.0.0
platformdirs==3.2.0
psycopg2==2.9.7
pymongo==4.3.3
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.21.0
Werkzeug==2.3.6
